alice_trading/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pya3==1.0.8
python-dotenv==1.0.0
pyotp==2.9.0
websocket-client==1.6.4
google-generativeai==0.8.6
requests==2.31.0
pydantic==2.5.1
